requirements.txt
mysql-connector-python==9.6.0
